landmark-tracking/requirements.txt
rclpy
sensor_msgs
opencv-python
opencv-contrib-python
requests
numpy
pillow
scikit-image
scipy
torch
matplotlib
cv-bridge